Library Reference

Core.Type — Method

WiringDiagram(inputs::AbstractVector, output::AbstractVector[, label::AbstractDict])

Construct a wiring diagram. The vector inputs specifies the function

port(b, i) := inputs[b][i]

for all boxes b ∈ B and i ∈ {1, ..., |box⁻¹(b)|}. The vector output specifies the function

outwire(q) := output[q]

for all outer ports q ∈ Q, and the dictionary label specifies the function

label(w) := label[w]

for all wires w ∈ W. If label is omitted, then the constructed wiring diagram will be unlabeled (i.e. all labels will be set to nothing).

WiringDiagrams.AbstractAlgebra — Type

AbstractAlgebra{A}

A wiring diagram algebra a is a higher-order function that transforms each wiring diagram d into a multi-function

a(d): X₁, ..., Xₙ → Y.

WiringDiagrams.AbstractDendrogram — Type

AbstractDendrogram{I, L}

A labelled dendrogram is a commutative dendrogram in Set of the form

+ -------------------- +
|         L label      |
|    wire ↑ outbox     |
|     P → W ← Q        |
| box ↓   ↓   ↓ outbox |
|     B → C ← C        |
|  parent  outparent   |
+ -------------------- +

where

B is a set of boxes

C is a set of outer boxes

W is a set of wires

P is a set of ports

Q is a set of outer ports

L is a set of labels

The functions parent and outparent define a rooted tree with leaves B and branches C.

+ --------------- +
|         c       |
|         ↓       |
|         c       |
|       ↙ ↓  ↘    |
|     c   c    b  |
|   ↙   ↘   ↘     |
|  b     b    b   |
+ --------------- +

The sets B, C, W, P, and Q are contiguous sets of natural numbers

B = {1, ..., |B|}

C = {1, ..., |C|}

W = {1, ..., |W|}

P = {1, ..., |P|}

Q = {1, ..., |Q|}

and the set L is an arbitrary Julia type like Symbol or Int. The functions

box: P → B

outbox: Q → C

wirebox: W → C

are monotonically increasing, to that for all boxes b ∈ B and outer boxes c ∈ C, the pre-images

box⁻¹(b) = {p ∈ P : box(p) = b} ⊆ P

outbox⁻¹(c) = {q ∈ Q : outbox(q) = c} ⊆ Q

wirebox⁻¹(c) = {w ∈ W : wirebox(w) = c} ⊆ W

are contiguos sets of natural numbers:

box⁻¹(b) = {p, ..., p + |box⁻¹(b)| - 1} ⊆ P

outbox⁻¹(c) = {q, ..., q + |outbox⁻¹(c)| - 1} ⊆ Q

wirebox⁻¹(c) = {w, ..., w + |wirebox⁻¹(c)| - 1} ⊆ W

Hence, there are functions port, outport, and boxwire which map each box b ∈ B, outer box c ∈ C, and numbers

1 ≤ i ≤ |box⁻¹(b)|

1 ≤ j ≤ |outbox⁻¹(c)|

1 ≤ k ≤ |wirebox⁻¹(c)|

to the objects

port(b, i) := p + i - 1 ∈ box⁻¹(b)

outport(q, j) := q + j - 1 ∈ outbox⁻¹(c)

boxwire(w, k) := w + k - 1 ∈ wirebox⁻¹(c)

WiringDiagrams.AbstractOperation — Type

AbstractOperation{I, L}

A morphism in a multi-category.

WiringDiagrams.AbstractWiringDiagram — Type

AbstractWiringDiagram{I, L} <: AbstractOperation{I, L}

A labelled wiring diagram is a diagram in Set of the form

+ --------------- +
|       B         |
|       ↑ box     |
|       P         |
| label ↓ wire    |
|   L ← W         |
|       ↑ outwire |
|       Q         |
+ --------------- +

where

B is a set of boxes

W is a set of wires

P is a set of ports

Q is a set of outer ports

L is a set of labels

The sets B, W, P, and Q are contiguous sets of natural numbers

B = {1, ..., |B|}

W = {1, ..., |W|}

P = {1, ..., |P|}

Q = {1, ..., |Q|}

and the set L is an arbitrary Julia type like Symbol or Int. The function box: P → Q is monotonically increasing, so that for all boxes b ∈ B, the pre-image

box⁻¹(b) = {p ∈ P : box(p) = b} ⊆ P

is a contiguous set of natural numbers

box⁻¹(b) = {p, ..., p + |box⁻¹(b)| - 1} ⊆ P.

Hence, there is a function port which maps each box b ∈ B and number 1 ≤ i ≤ |box⁻¹(b)| to the port

port(b, i) := p + i - 1 ∈ box⁻¹(b).

WiringDiagrams.Dendrogram — Type

Dendrogram{I, L, Xnp, Lbl, Wre, PrtLbl, Out, OutPrtLbl} <: AbstractDendrogram{I, L}

WiringDiagrams.Operation — Type

Operation{I, L, Alg, Dgm} <: AbstractOperation{I, L}

A multi-function

f: X₁, X₂, ..., Xₙ → Y

of the form

f(x₁, x₂, ..., xₙ) = a(d)(x₁, x₂, ..., xₙ)

for some wiring diagram algebra a and wiring diagram d.

WiringDiagrams.WiringDiagram — Type

WiringDiagram{I, L, XPrt, Lbl, Wre, PrtLbl, Out, OutPrtLbl} <: AbstractWiringDiagram{I, L}

WiringDiagrams.WiringDiagram — Method

WiringDiagram{I, L}(nb::Integer, nw::Integer, np::Integer, nop::Integer) where {I <: Integer, L}

Construct an uninitialized diagram with dimensions

|B| = nb
|W| = nw
|P| = np
|O| = nop.

AbstractTrees.parent — Method

parent(dendrogram::AbstractDendrogram)

Get the parent of a box b ∈ B:

parent(b) ∈ C.

WiringDiagrams.apply — Method

AbstractAlgebra{A}

Given a wiring diagram algebra a, a wiring diagram d, and a sequence

(x₁, x₂, ..., xₙ)

of arguments, compute y:

y = a(d)(x₁, x₂, ..., xₙ).

WiringDiagrams.arity — Method

arity(operation::AbstractOperation)

Get the arity of an operation.

WiringDiagrams.boxes — Method

boxes(dendrogram::AbstractDendrogram, c::Integer)

Get the ordered pre-image parent⁻¹(c) ⊆ B:

{b ∈ B : parent(b) = c} ⊆ B.

WiringDiagrams.boxes — Method

boxes(dendrogram::AbstractDendrogram)

Get the ordered set of boxes in a dendrogram:

B.

WiringDiagrams.boxes — Method

boxes(diagram::AbstractWiringDiagram)

Get the ordered set of boxes in a wiring diagram:

B.

WiringDiagrams.codomain — Method

codomain(operation::AbstractOperation)

Get the codomain of an operation.

WiringDiagrams.compose — Method

compose(i::Integer, outer::AbstractOperation, inner::AbstractOperation)

Given operations

inner: (A₁, …, Aₘ) → Bᵢ
outer: (B₁, …, Bₙ) → C

form the composite operation

outer ∘ᵢ inner: (B₁, …, Bᵢ₋₁, A₁, …, Aₘ, Bᵢ₊₁, …, Bₙ) → C

given by the following string diagram:

                  ┌───────┐
B₁ ───────────────│       │
 ⋮                │       │
Bᵢ₋₁ ─────────────│       │
      ┌───────┐   │       │
A₁ ───│       │ Bᵢ│       │
 ⋮    │ inner │───│ outer │─── C
Aₘ ───│       │   │       │
      └───────┘   │       │
Bᵢ₊₁ ─────────────│       │
 ⋮                │       │
Bₙ ───────────────│       │
                  └───────┘

WiringDiagrams.domain — Method

domain(operation::AbstractOperation, i::Integer)

Get the ith element of the domain of an operation.

WiringDiagrams.domain — Method

domain(operation::AbstractOperation)

Get the domain of an operation.

WiringDiagrams.eachargument — Method

eachargument(operation::AbstractOperation)

Get the ordered set

{1, ..., n} ⊆ ℕ

where n is the arity of the operation.

WiringDiagrams.label — Method

label(dendrogram::AbstractDendrogram, w::Integer)

Get the label of a wire w ∈ W:

label(w) ∈ L.

WiringDiagrams.label — Method

label(diagram::AbstractWiringDiagram, w::Integer)

Get the label of a wire w ∈ W:

label(w) ∈ L.

WiringDiagrams.nb — Method

nb(dendrogram::AbstractDendrogram, c::Integer)

Get the size of the pre-image parent⁻¹(c) ⊆ B:

| {b ∈ B : parent(b) = c} |.

WiringDiagrams.nb — Method

nb(dendrogram::AbstractDendrogram)

Get the number of boxes in a dendrogram:

|B|.

WiringDiagrams.nb — Method

nb(diagram::AbstractWiringDiagram)

Get the number of boxes in a wiring diagram:

|B|.

WiringDiagrams.nob — Method

nob(dendrogram::AbstractDendrogram, c::Integer)

Get the size of the pre-image outparent⁻¹(c) ⊆ C:

| {cc ∈ C : outparent(cc) = c} |.

WiringDiagrams.nob — Method

nob(dendrogram::AbstractDendrogram)

Get the number of outer boxes in a dendrogram:

|C|.

WiringDiagrams.nop — Method

nop(dendrogram::AbstractDendrogram, c::Integer)

Get the size of the pre-image outbox⁻¹(c) ⊆ Q:

| {q ∈ Q : outbox(q) = c} |.

WiringDiagrams.nop — Method

nop(dendrogram::AbstractDendrogram)

Get the number of outer ports in a dendrogram:

|Q|.

WiringDiagrams.nop — Method

nop(diagram::AbstractWiringDiagram)

Get the number of outer ports in a wiring diagram:

|Q|.

WiringDiagrams.np — Method

ports(dendrogram::AbstractDendrogram, b::Integer)

Get the size of the pre-image box⁻¹(b) ⊆ P:

| {p ∈ P : box(p) = b} |.

WiringDiagrams.np — Method

np(dendrogram::AbstractDendrogram)

Get the number of ports in a dendrogram:

|P|.

WiringDiagrams.np — Method

np(diagram::AbstractWiringDiagram, b::Integer)

Get the size of the preimage box⁻¹(b) ⊆ P:

| {p ∈ P : box(p) = b} |.

WiringDiagrams.np — Method

np(diagram::AbstractWiringDiagram)

Get the number of ports in a wiring diagram:

|P|.

WiringDiagrams.nw — Method

nw(dendrogram::AbstractDendrogram, c::Integer)

Get the size of the pre-image wirebox⁻¹(c) ⊆ W:

| {w ∈ W : wirebox(w) = c} |.

WiringDiagrams.nw — Method

nw(dendrogram::AbstractDendrogram)

Get the number of wires in a dendrogram:

|W|.

WiringDiagrams.nw — Method

nw(diagram::AbstractWiringDiagram)

Get the number of wires in a wiring diagram.

|W|.

WiringDiagrams.outboxes — Method

outboxes(dendrogram::AbstractDendrogram, c::Integer)

Get the ordered pre-image outparent⁻¹(c) ⊆ C:

{cc ∈ C : outparent(cc) = c} ⊆ C.

WiringDiagrams.outboxes — Method

outboxes(dendrogram::AbstractDendrogram)

Get the orderd set of outer boxes in a dendrogram:

C.

WiringDiagrams.outparent — Method

outparent(dendrogram::AbstractDendrogram, c::Integer)

Get the parent of an outer box c ∈ C:

outparent(c) ∈ C.

WiringDiagrams.outportlabel — Method

outportlabel(dendrogram::AbstractDendrogram, q::Integer)

Get the label of the wire of an outer port q ∈ Q:

label(outwire(q)) ∈ L.

WiringDiagrams.outportlabel — Method

outportlabel(diagram::AbstractWiringDiagram, q::Integer)

Get the label of the wire of an outer port q ∈ Q:

label(outwire(q)) ∈ L.

WiringDiagrams.outportlabels — Method

outportlabels(dendrogram::AbstractDendrogram, c::Integer)

Get the composite function {1, ..., |outbox⁻¹(c)|} → L:

outportlabels(i) := label(outwire(outport(c, i))).

WiringDiagrams.outportlabels — Method

outlabels(diagram::AbstractWiringDiagram)

Get the composite function Q → L:

outlabels(q) := label(outwire(q))

WiringDiagrams.outports — Method

outports(dendrogram::AbstractDendrogram, c::Integer)

Get the ordered pre-image outbox⁻¹(c) ⊆ Q:

{q ∈ Q : outbox(q) = c} ⊆ Q.

WiringDiagrams.outports — Method

outports(dendrogram::AbstractDendrogram)

Get the ordered set of outer ports in a dendrogram:

Q.

WiringDiagrams.outports — Method

outports(diagram::AbstractWiringDiagram)

Get the ordered set of outer ports in a wiring diagram:

Q.

WiringDiagrams.outportwires — Method

outportwires(dendrogram::AbstractDendrogram, c::Integer)

Get the composite function {1, ..., |outbox⁻¹(c)|} → W:

outportwires(i) := outwire(outport(c, i)).

WiringDiagrams.outportwires — Method

outportwires(diagram::AbstractWiringDiagram)

Get the function Q → W:

outportwires(q) := outwire(q)

WiringDiagrams.outwire — Method

outwire(dendrogram::AbstractDendrogram, q::Integer)

Get the wire of an outer port q ∈ O:

outwire(q) ∈ W.

WiringDiagrams.outwire — Method

outwire(diagram::AbstractWiringDiagram, q::Integer)

Get the wire of an outer port q ∈ O:

outwire(q) ∈ W.

WiringDiagrams.portlabel — Method

portlabel(dendrogram::AbstractDendrogram, p::Integer)

Get the label of the wire of a port p ∈ P:

label(wire(p)) ∈ L.

WiringDiagrams.portlabel — Method

portlabel(diagram::AbstractWiringDiagram, p::Integer)

Get the label of the wire of a port p ∈ P:

label(wire(p)) ∈ L.

WiringDiagrams.portlabels — Method

portlabels(dendrogram::AbstractDendrogram, b::Integer)

Get the composite function {1, ..., |box⁻¹(b)|} → L:

portlabels(i) := label(wire(port(b, i))).

WiringDiagrams.portlabels — Method

portlabels(diagram::AbstractWiringDiagram, b::Integer)

Get the composite function {1, ..., |box⁻¹(b)|} → L:

portlabels(i) := label(wire(port(b, i))).

WiringDiagrams.ports — Method

ports(dendrogram::AbstractDendrogram, b::Integer)

Get the ordered pre-image box⁻¹(b) ⊆ P:

{p ∈ P : box(p) = b} ⊆ P.

WiringDiagrams.ports — Method

ports(dendrogram::AbstractDendrogram)

Get the ordered set of ports in a dendrogram:

P.

WiringDiagrams.ports — Method

ports(diagram::AbstractWiringDiagram, b::Integer)

Get the ordered preimage box⁻¹(b) ⊆ P:

{p ∈ P : box(p) = b} ⊆ P.

WiringDiagrams.ports — Method

ports(diagram::AbstractWiringDiagram)

Get the ordered set of ports in a wiring diagram:

P.

WiringDiagrams.portwires — Method

portwires(dendrogram::AbstractDendrogram, b::Integer)

Get the composite function {1, ..., |box⁻¹(b)|} → W:

portwires(i) := wire(port(b, i)).

WiringDiagrams.portwires — Method

portwires(diagram::AbstractWiringDiagram, b::Integer)

Get the composite function {1, ..., |box⁻¹(b)|} → W:

portwires(i) := wire(port(b, i)).

WiringDiagrams.portwires — Method

portwires(diagram::AbstractWiringDiagram)

Get the function P → W:

portwires(p) := wire(p)

WiringDiagrams.wire — Method

wire(dendrogram::AbstractDendrogram, p::Integer)

Get the wire of a port p ∈ P:

wire(p) ∈ W.

WiringDiagrams.wire — Method

wire(diagram::AbstractWiringDiagram, p::Integer)

Get the wire of a port p ∈ P:

wire(p) ∈ W.

WiringDiagrams.wirelabels — Method

wirelabels(dendrogram::AbstractDendrogram, c::Integer)

Get the composite function {1, ..., |wirebox⁻¹(c)|} → L:

wirelabels(i) := label(boxwire(c, i))

WiringDiagrams.wirelabels — Method

wirelabels(dendrogram::AbstractDendrogram)

Get the function of W → L:

wirelabels(w) := label(w).

WiringDiagrams.wirelabels — Method

wirelabels(diagram::AbstractWiringDiagram)

Get the function W → L:

wirelabels(w) := label(w)

WiringDiagrams.wires — Method

wires(dendrogram::AbstractDendrogram, c::Integer)

Get the ordered pre-image wirebox⁻¹(c) ⊆ W:

{w ∈ W : wirebox(w) = c} ⊆ W.

WiringDiagrams.wires — Method

wires(dendrogram::AbstractDendrogram)

Get the ordered set of wires in a dendrogram:

W.

WiringDiagrams.wires — Method

wires(diagram::AbstractWiringDiagram)

Get the ordered set of wires in a wiring diagram:

W.